backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.15.2",
    "bcrypt>=4.3.0",
    "fastapi[standard]>=0.115.12",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.9.1",
    "pyjwt>=2.10.1",
    "sqlalchemy>=2.0.40",
]
